Move locked figure tests
