+++ packages/perseus-editor/src/components/__stories__/locked-function-settings.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import LockedFunctionSettings from "../graph-locked-figures/locked-function-settings";
+import LockedFunctionSettings from "../../widgets/interactive-graph-editor/locked-figures/locked-function-settings";
 import {getDefaultFigureForType} from "../util";
 
 import type {Meta, StoryObj} from "@storybook/react";
