+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/color-select.tsx
