+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/ellipse-swatch.tsx
