+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/labeled-switch.tsx
