+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/line-stroke-select.tsx
