+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/locked-figure-select.tsx
