+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/locked-figures-section.test.tsx
@@ -3,8 +3,8 @@ import {screen, render} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import LockedFiguresSection from "../graph-locked-figures/locked-figures-section";
-import {getDefaultFigureForType} from "../util";
+import LockedFiguresSection from "./locked-figures-section";
+import {getDefaultFigureForType} from "../../../components/util";
 
 import type {UserEvent} from "@testing-library/user-event";
