+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/locked-function-examples.ts
