+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/polygon-swatch.tsx
